System.Console.MinTTY

Description

Exports functions that check if a process or handle is attached to a MinTTY console on Windows, such as Cygwin or MSYS. On non-Windows operating systems, the functions in this module will simply return False.

Documentation

isMinTTY :: IO Bool

Returns True if the current process's standard error is attached to a MinTTY console (e.g., Cygwin or MSYS). Returns False otherwise.

isMinTTYHandle :: HANDLE -> IO Bool

Returns True is the given handle is attached to a MinTTY console (e.g., Cygwin or MSYS). Returns False otherwise.

type HANDLE = Ptr ()